editoast: add track occupancy for paced train endpoint
