+++ editoast/src/views/timetable.rs
@@ -18,6 +18,7 @@ use axum::extract::State;
 use axum::http::StatusCode;
 use axum::response::IntoResponse;
 use chrono::DateTime;
+use chrono::Duration;
 use chrono::Utc;
 use core_client::AsCoreRequest;
 use core_client::CoreClient;
@@ -27,6 +28,7 @@ use core_client::conflict_detection::ConflictRequirement;
 use core_client::conflict_detection::ConflictType;
 use core_client::conflict_detection::TrainRequirements;
 use core_client::simulation::PhysicsConsist;
+use core_client::simulation::ReportTrain;
 use editoast_authz as authz;
 use editoast_common::units::quantities::Acceleration;
 use editoast_common::units::quantities::Length;
@@ -35,11 +37,15 @@ use editoast_common::units::quantities::Velocity;
 use editoast_derive::EditoastError;
 use editoast_models::DbConnection;
 use editoast_models::DbConnectionPoolV2;
+use editoast_schemas::infra::TrackOffset;
 use editoast_schemas::paced_train::PacedTrain;
+use editoast_schemas::primitives::PositiveDuration;
 use editoast_schemas::rolling_stock::EtcsBrakeParams;
 use editoast_schemas::rolling_stock::RollingResistance;
 use editoast_schemas::rolling_stock::RollingStock;
 use editoast_schemas::rolling_stock::TowedRollingStock;
+use editoast_schemas::train_schedule::OperationalPointIdentifier;
+use editoast_schemas::train_schedule::PathItemLocation;
 use editoast_schemas::train_schedule::TrainSchedule;
 use itertools::Either;
 use itertools::Itertools;
@@ -63,6 +69,7 @@ use crate::ValkeyClient;
 use crate::error::Result;
 use crate::models;
 use crate::models::Infra;
+use crate::models::OperationalPointModel;
 use crate::models::paced_train::PacedTrainChangeset;
 use crate::models::prelude::*;
 use crate::models::timetable::Timetable;
@@ -70,6 +77,11 @@ use crate::models::timetable::TimetableWithTrains;
 use crate::models::train_schedule::TrainScheduleChangeset;
 use crate::views::AuthenticationExt;
 use crate::views::AuthorizationError;
+use crate::views::path::path_item_cache::PathItemCache;
+use crate::views::path::projection::PathProjection;
+use crate::views::path::projection::TrackLocationFromPath;
+use crate::views::projection::find_index_upper;
+use crate::views::projection::interpolate;
 
 crate::routes! {
     "/timetable" => {
@@ -119,6 +131,9 @@ enum TimetableError {
     #[error("Failed to parse train_id '{train_id}'")]
     #[editoast_error(status = 500)]
     ParseError { train_id: String },
+    #[error("Operational point '{operational_point_id}', could not be found")]
+    #[editoast_error(status = 404)]
+    OperationalPointNotFound { operational_point_id: String },
 }
 
 /// Creation result for a Timetable
@@ -786,6 +801,291 @@ fn build_conflict_core_request(
     }
 }
 
+#[derive(Debug, Clone, Serialize, Deserialize, ToSchema)]
+#[schema(as = TrainScheduleTrackOccupancy)]
+pub struct TrackOccupancy {
+    time_begin: DateTime<Utc>,
+    #[schema(value_type = chrono::Duration, example = "PT5M")]
+    duration: PositiveDuration,
+    train_schedule_id: i64,
+}
+
+/// Match the id of an operational point with a path item in the train schedule and return the id of the path item concerned
+fn match_path_item_id_with_op_id<'a>(
+    path_item_cache: &PathItemCache,
+    train_schedule: &'a models::TrainSchedule,
+    operational_point_id: &str,
+) -> Option<&'a str> {
+    for path_item in &train_schedule.path {
+        if let PathItemLocation::OperationalPointReference(operational_point_reference) =
+            &path_item.location
+        {
+            match &operational_point_reference.reference {
+                OperationalPointIdentifier::OperationalPointId { operational_point } => {
+                    if operational_point.0 == operational_point_id {
+                        return Some(&path_item.id);
+                    }
+                }
+                OperationalPointIdentifier::OperationalPointDescription { trigram, .. } => {
+                    let operational_points = path_item_cache
+                        .get_from_trigram(&trigram.0)
+                        .expect("The operational points are supposed to exist");
+                    if operational_points
+                        .iter()
+                        .any(|op| op.obj_id == operational_point_id)
+                    {
+                        return Some(&path_item.id);
+                    }
+                }
+                OperationalPointIdentifier::OperationalPointUic { uic, .. } => {
+                    let ops = path_item_cache
+                        .get_from_uic(*uic as i64)
+                        .expect("The operational points are supposed to exist");
+                    if ops.iter().any(|op| op.obj_id == operational_point_id) {
+                        return Some(&path_item.id);
+                    }
+                }
+            }
+        }
+    }
+    None
+}
+
+/// Computes the occupancy of track sections for a given train based on its schedule and path.
+/// If a path item ID is provided, it uses schedule data to determine stop duration and location.
+fn track_occupancy_on_path_item(
+    train_schedule: &models::TrainSchedule,
+    path_item_id: &str,
+    path_item_positions: &[u64],
+    path_projection: &PathProjection,
+    operational_point_track_offsets: &[TrackOffset],
+    report_train: &ReportTrain,
+) -> Option<(String, TrackOccupancy)> {
+    let schedule_item = train_schedule
+        .schedule
+        .iter()
+        .find(|schedule| schedule.at.0 == path_item_id);
+
+    // Stop time at the operational point:
+    // Retrieve the stop time if it exists; otherwise, default to 0
+    // the path_item is not necessarily a schedule point.
+    // if it's not a schedule point, the stop_for is 0. (PositiveDuration::default())
+    let stop_duration = schedule_item
+        .and_then(|item| item.stop_for.clone())
+        .unwrap_or_default();
+
+    // Get the index of the path item in the list of path items that corresponds to the operational point
+    let index = train_schedule
+        .path
+        .iter()
+        .position(|path| path.id.0 == path_item_id)
+        .expect("Unexpected path item");
+
+    let position = path_item_positions[index];
+    let time = report_train.path_item_times[index];
+
+    // Retrieve the track section that matches the path item's position along the path
+    let track_section = match path_projection.get_location(position) {
+        TrackLocationFromPath::One(track_offset) => operational_point_track_offsets
+            .iter()
+            .find(|to| *to == &track_offset)
+            .expect("Invalid track offset")
+            .track
+            .to_string(),
+        TrackLocationFromPath::Two(track_offset, track_offset1) => operational_point_track_offsets
+            .iter()
+            .find(|to| *to == &track_offset)
+            .or(operational_point_track_offsets
+                .iter()
+                .find(|to| *to == &track_offset1))
+            .expect("Invalid track offset")
+            .track
+            .to_string(),
+    };
+
+    Some((
+        track_section,
+        TrackOccupancy {
+            time_begin: train_schedule.start_time + Duration::milliseconds(time as i64),
+            duration: stop_duration,
+            train_schedule_id: train_schedule.id,
+        },
+    ))
+}
+
+pub fn find_track_occupancy_for_operational_point(
+    operational_point_id: &str,
+    operational_point_track_offsets: &[TrackOffset],
+    path_item_cache: &PathItemCache,
+    simulation: &simulation::Response,
+    pathfinding: &PathfindingResult,
+    train_schedule: &models::TrainSchedule,
+) -> Vec<(String, TrackOccupancy)> {
+    // Get the track ranges from the pathfinding
+
+    let mut track_occupancies = vec![];
+    let (track_ranges, path_item_positions) = match pathfinding {
+        PathfindingResult::Success(pathfinding_result_success) => {
+            assert_eq!(
+                pathfinding_result_success.path_item_positions.len(),
+                train_schedule.path.len()
+            );
+            (
+                &pathfinding_result_success.track_section_ranges,
+                &pathfinding_result_success.path_item_positions,
+            )
+        }
+        PathfindingResult::Failure(_) => {
+            // The timetable can have a pathfinding failure for a train schedule
+            // because the train schedule is not valid.
+            // We skip it because we don't want to compute the track occupancy for an invalid train schedule.
+            // It's not a problem because we are just looking for the track occupancy of the operational point for the valid trains.
+            tracing::info!(train_schedule.id, "pathfinding failed");
+            return track_occupancies;
+        }
+    };
+    let path_projection = PathProjection::new(track_ranges);
+
+    // Get the positions and the times from the simulation
+    let report_train = match simulation {
+        simulation::Response::Success { final_output, .. } => &final_output.report_train,
+        _ => {
+            tracing::info!(train_schedule.id, "simulation failed");
+            return track_occupancies;
+        }
+    };
+
+    let path_item_id =
+        match_path_item_id_with_op_id(path_item_cache, train_schedule, operational_point_id);
+    if let Some(path_item_id) = path_item_id {
+        if let Some(track_occupancy) = track_occupancy_on_path_item(
+            train_schedule,
+            path_item_id,
+            path_item_positions,
+            &path_projection,
+            operational_point_track_offsets,
+            report_train,
+        ) {
+            track_occupancies.push(track_occupancy);
+        }
+    } else {
+        track_occupancies.extend(interpolate_track_occupancy(
+            train_schedule,
+            &path_projection,
+            operational_point_track_offsets,
+            report_train,
+        ));
+    }
+    track_occupancies
+}
+
+fn interpolate_track_occupancy(
+    train_schedule: &models::TrainSchedule,
+    path_projection: &PathProjection,
+    operational_point_track_offsets: &[TrackOffset],
+    report_train: &ReportTrain,
+) -> Vec<(String, TrackOccupancy)> {
+    // Iterate over the track offsets linked to the operational point
+    // until we find its corresponding positions on the pathfinding route.
+    // Then, retrieve the associated times (via interpolation) and positions on the path.
+    operational_point_track_offsets
+        .iter()
+        .filter_map(|track_offset| {
+            path_projection.get_position(track_offset).map(|position| {
+                let index = find_index_upper(&report_train.positions, position);
+                let time = if index == 0 {
+                    report_train.times[0]
+                } else {
+                    interpolate(
+                        report_train.positions[index - 1],
+                        report_train.positions[index],
+                        report_train.times[index - 1],
+                        report_train.times[index],
+                        position,
+                    )
+                };
+                (
+                    track_offset.track.to_string(),
+                    TrackOccupancy {
+                        time_begin: train_schedule.start_time + Duration::milliseconds(time as i64),
+                        duration: PositiveDuration::default(),
+                        train_schedule_id: train_schedule.id,
+                    },
+                )
+            })
+        })
+        .collect()
+}
+
+async fn compute_track_occupancy(
+    db_pool: Arc<DbConnectionPoolV2>,
+    valkey: Arc<ValkeyClient>,
+    core_client: Arc<CoreClient>,
+    operational_point_id: String,
+    train_schedules: &[models::TrainSchedule],
+    infra_id: i64,
+    electrical_profile_set_id: Option<i64>,
+) -> Result<HashMap<String, Vec<TrackOccupancy>>> {
+    let infra = Infra::retrieve_real_or_fail(db_pool.get().await?, infra_id, || {
+        TimetableError::InfraNotFound { infra_id }
+    })
+    .await?;
+
+    let operational_point = OperationalPointModel::retrieve_real_or_fail(
+        db_pool.get().await?,
+        (infra_id, operational_point_id.clone()),
+        || TimetableError::OperationalPointNotFound {
+            operational_point_id: operational_point_id.clone(),
+        },
+    )
+    .await?;
+
+    // Retrieve simulations
+    let simulations_result = train_simulation_batch(
+        &mut db_pool.get().await?,
+        valkey,
+        core_client,
+        &train_schedules,
+        &infra,
+        electrical_profile_set_id,
+    )
+    .await?;
+    let operational_point_track_offsets = operational_point.schema.track_offset();
+
+    let path_items = train_schedules
+        .iter()
+        .flat_map(|ts| ts.path.iter().map(|p| &p.location))
+        .collect::<Vec<_>>();
+
+    let path_item_cache =
+        PathItemCache::load(&mut db_pool.get().await?, infra.id, &path_items).await?;
+
+    let track_occupancy_map = simulations_result
+        .iter()
+        .zip(train_schedules)
+        .flat_map(|((simulation, pathfinding), train_schedule)| {
+            find_track_occupancy_for_operational_point(
+                &operational_point_id,
+                &operational_point_track_offsets,
+                &path_item_cache,
+                simulation,
+                pathfinding,
+                &train_schedule,
+            )
+        })
+        .fold(
+            HashMap::<String, Vec<TrackOccupancy>>::new(),
+            |mut track_occupancy_map, (track_section, track_occupancy)| {
+                track_occupancy_map
+                    .entry(track_section)
+                    .or_default()
+                    .push(track_occupancy);
+                track_occupancy_map
+            },
+        );
+    Ok(track_occupancy_map)
+}
+
 #[derive(Debug, Clone)]
 pub struct PhysicsConsistParameters {
     pub total_mass: Option<Mass>,
@@ -988,6 +1288,7 @@ mod tests {
     use editoast_schemas::fixtures::simple_modified_exception_with_change_groups;
     use editoast_schemas::fixtures::simple_rolling_stock;
     use editoast_schemas::fixtures::towed_rolling_stock;
+    use editoast_schemas::infra::Direction;
     use editoast_schemas::paced_train::ExceptionType;
     use editoast_schemas::paced_train::PacedTrainException;
     use editoast_schemas::paced_train::PathAndScheduleChangeGroup;
@@ -1577,6 +1878,186 @@ mod tests {
         assert_eq!(
             physics_consist.compute_const_gamma(),
             units::meter_per_second_squared::new(0.6)
+        );
+    }
+
+    #[rstest]
+    #[case("T2", 50, Direction::StartToStop, vec![0, 100], vec![0, 1000], Some("T2"))]
+    #[case("T2", 150, Direction::StartToStop, vec![0, 100], vec![0, 1000], None)]
+    #[case("T2", 50, Direction::StopToStart, vec![0, 100], vec![0, 1000], Some("T2"))]
+    fn test_interpolate_track_occupancy(
+        #[case] track_name: &str,
+        #[case] offset: u64,
+        #[case] direction: Direction,
+        #[case] positions: Vec<u64>,
+        #[case] times: Vec<u64>,
+        #[case] expected_track: Option<&str>,
+    ) {
+        use core_client::pathfinding::TrackRange;
+        use editoast_schemas::primitives::Identifier;
+
+        let track_section = Identifier::from(track_name);
+        let track_range = TrackRange {
+            track_section: track_section.clone(),
+            begin: 0,
+            end: 100,
+            direction,
+        };
+        let track_range_vec = vec![track_range];
+        let path_projection = PathProjection::new(&track_range_vec);
+        let operational_point_track_offsets = vec![TrackOffset {
+            track: track_section.clone(),
+            offset,
+        }];
+        let report_train = ReportTrain {
+            positions: positions.clone(),
+            times: times.clone(),
+            speeds: vec![10.0; positions.len()],
+            energy_consumption: 0.0,
+            path_item_times: vec![0, 1000],
+        };
+        let train_schedule = models::TrainSchedule::default();
+
+        let result = interpolate_track_occupancy(
+            &train_schedule,
+            &path_projection,
+            &operational_point_track_offsets,
+            &report_train,
+        );
+
+        assert_eq!(
+            result.first().map(|(track, _)| track.as_str()),
+            expected_track
+        );
+    }
+
+    #[rstest]
+    #[case("path_item_1", vec![50, 100, 150], 50, "T1", "T1", 0, 500)]
+    #[case("path_item_2", vec![50, 75, 150], 75, "T2", "T2", 1000, 300000)]
+    fn test_track_occupancy_on_path_item(
+        #[case] path_item_id: &str,
+        #[case] path_item_positions: Vec<u64>,
+        #[case] track_offset: u64,
+        #[case] track_name: &str,
+        #[case] expected_track: &str,
+        #[case] expected_time: u64,
+        #[case] expected_stop_duration_ms: i64,
+    ) {
+        // Create test data
+
+        use core_client::pathfinding::TrackRange;
+        use editoast_schemas::{
+            infra::Direction,
+            primitives::Identifier,
+            train_schedule::{OperationalPointReference, PathItem, ReceptionSignal, ScheduleItem},
+        };
+        let track_section = Identifier::from(track_name);
+        let track_range = TrackRange {
+            track_section: track_section.clone(),
+            begin: 0,
+            end: 150,
+            direction: Direction::StartToStop,
+        };
+        let track_ranges = vec![track_range];
+        let path_projection = PathProjection::new(&track_ranges);
+
+        let operational_point_track_offsets = vec![TrackOffset {
+            track: track_section.clone(),
+            offset: track_offset,
+        }];
+
+        let report_train = ReportTrain {
+            positions: vec![0, 50, 100, 150],
+            times: vec![0, expected_time, 2000, 3000],
+            speeds: vec![10.0; 4],
+            energy_consumption: 0.0,
+            path_item_times: vec![0, expected_time, 2000],
+        };
+
+        // Create a train schedule with path items and schedule items
+        let train_schedule = models::TrainSchedule {
+            start_time: DateTime::from_timestamp(1000000000, 0).unwrap(),
+            path: vec![
+                PathItem {
+                    id: "path_item_1".into(),
+                    location: PathItemLocation::OperationalPointReference(
+                        OperationalPointReference {
+                            reference: OperationalPointIdentifier::OperationalPointId {
+                                operational_point: "op_1".into(),
+                            },
+                            track_reference: None,
+                        },
+                    ),
+                    deleted: false,
+                },
+                PathItem {
+                    id: "path_item_2".into(),
+                    location: PathItemLocation::OperationalPointReference(
+                        OperationalPointReference {
+                            reference: OperationalPointIdentifier::OperationalPointId {
+                                operational_point: "op_2".into(),
+                            },
+                            track_reference: None,
+                        },
+                    ),
+                    deleted: false,
+                },
+                PathItem {
+                    id: "path_item_invalid".into(),
+                    location: PathItemLocation::OperationalPointReference(
+                        OperationalPointReference {
+                            reference: OperationalPointIdentifier::OperationalPointId {
+                                operational_point: "op_invalid".into(),
+                            },
+                            track_reference: None,
+                        },
+                    ),
+                    deleted: false,
+                },
+            ],
+            schedule: vec![
+                ScheduleItem {
+                    at: "path_item_1".into(),
+                    arrival: None,
+                    stop_for: Some(
+                        PositiveDuration::try_from(Duration::milliseconds(500)).unwrap(),
+                    ),
+                    locked: false,
+                    reception_signal: ReceptionSignal::Open,
+                },
+                ScheduleItem {
+                    at: "path_item_2".into(),
+                    arrival: None,
+                    stop_for: Some(
+                        PositiveDuration::try_from(Duration::milliseconds(300000)).unwrap(),
+                    ),
+                    locked: false,
+                    reception_signal: ReceptionSignal::Open,
+                },
+            ],
+            ..models::TrainSchedule::default()
+        };
+
+        // Call the function
+        let (track, occupancy) = track_occupancy_on_path_item(
+            &train_schedule,
+            path_item_id,
+            &path_item_positions,
+            &path_projection,
+            &operational_point_track_offsets,
+            &report_train,
+        )
+        .expect("computation should be successful according to test inputs");
+
+        assert_eq!(track, expected_track);
+        assert_eq!(occupancy.train_schedule_id, train_schedule.id);
+        assert_eq!(
+            occupancy.time_begin,
+            train_schedule.start_time + Duration::milliseconds(expected_time as i64)
+        );
+        assert_eq!(
+            occupancy.duration,
+            PositiveDuration::try_from(Duration::milliseconds(expected_stop_duration_ms)).unwrap()
         );
     }
 }
